feat: Account 도메인 시큐리티 및 예외 처리 구현

+++ Backend/src/main/java/gwangju/ssafy/backend/domain/account/repository/TransactionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TransactionRepository extends JpaRepository<Transaction, Long>,TransactionBatchRepository, QueryDslTransactionRepository {
 
-	int countByGroupAccount_Id(Long groupAccountId);
+	Long countByGroupAccount_Id(Long groupAccountId);
 
 }
